Refactoring search logic to use adapter pattern
